Share ShareAttachmentFlags
