+++ MWSE/xActivate.cpp
@@ -71,13 +71,13 @@ namespace mwse {
 			activator->type = TES3::AttachmentType::Action;
 			activator->next = nullptr;
 			activator->reference = nullptr;
-			activator->flags = 1;
+			activator->flags = TES3::ActionFlags::UseEnabled;
 
 			target->insertAttachment(activator);
 		}
 
 		// Ensure that our flags and reference is accurate.
-		activator->flags = 2;
+		activator->flags = TES3::ActionFlags::OnActivate;
 		activator->reference = reference;
 
 		// Call the original activate function.
