+++ SharedSE/SharedSE.vcxitems.filters
@@ -357,6 +357,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3SpellFlags.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3AttachmentFlags.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
